Improved my tests. Commented out the other because they're failing anyway
